Moved_ctrl_and_logic
