+++ server/controller/Articles.js
@@ -1,0 +1,33 @@
+module.exports = {
+  art_onload: async (req, res) => {
+    console.log(req);
+    try {
+      let db = req.app.get("db");
+      let result = await db.bpn_junction.test_bpn_query();
+      res.status(200).send(result);      
+    } catch (error) {
+      console.log(error);
+    }
+  },
+  art_topics :  async (req, res) => {
+    try {
+      let db = req.app.get("db");
+      let { topics }=req.params
+      let result = await db.articles.query_topics(topics)
+      res.status(200).send(result);
+    } catch (error) {
+      console.log(error)
+    }
+  },
+  get_topics: async (req, res) => {
+    try {
+      let db = req.app.get("db");
+      let result = await db.topics.get_all_topics();
+      res.status(200).send(result);      
+    } catch (error) {
+      console.log(error)
+    }
+  },
+
+
+}
